Create planner microservice (#20)
This PR creates and deploys the Planner micro-service: Use this to get
the time slots where users in a group are available. Note that this
returns every time slot where there's a change in availability. In the
future, we can filter options where less than x% of the group can meet
or where the duration is less than y minutes.

Please read the detailed documentation I included in the README! It has
everything you need to know to use this service in the apps :)